+++ python/lsst/ts/m1m3/utils/__init__.py
@@ -28,6 +28,7 @@ from .bump_test_runner import (
     ForceActuatorBumpTest,
 )
 from .bump_test_times import BumpTestTimes
+from .duration_time import DurationTime, parse_duration
 from .force_actuator_forces import ForceActuatorForces
 from .force_calculator import ForceCalculator
 from .simulator import Simulator